refactor: recruit 폴더명 project로 수정
